Enhanced reseeding of PRNGs
This commit makes the following changes:
1. When suhosin.mt_srand.ignore and/or suhosin.srand.ignore are enabled,
corresponding PRNGs are reseeded with random values (ignoring supplied
seeds) when mt_srand($seed) and/or srand($seed) are called. This way
applications that reseed PRNG(s) to create a brand new sequence of
random numbers will create a brand new sequence, but (in most cases)
with much more random/secure seed. Previous behavior was to ignore calls
to mt_srand() and srand() when suhosin.mt_srand.ignore and/or
suhosin.srand.ignore are enabled.
2. mt_rand() and rand() PRNGs are reseeded before first use in each new
request. This way it is much harder to predict random numbers generated
in past (or subsequest) requests that are server by the same PHP
process. Previous behaviour was to automatically seed PRNGs before first
use in each new PHP process.
3. A new php.ini value suhosin.rands.seedkey is introduced. This (along
with collected entropy) is cryptographically hashed and used as random
seeds for PRNGs. Setting this key to a long random and secret string
makes it even harder to predict generation of random PRNG seeds. This is
especially useful on systems without /dev/urandom.